Comitting a few attempts to speed up plotting with more concurrency. Still not great, since QtConcurrent::run() can't be interrupted, so quitting between dispatching and actual running will most likely cause segfaults. Could put try/catch around it...

+++ files.h
@@ -74,7 +74,8 @@ class DataFile : public QFile {
 		DataFile(QString &filename, QObject *parent = 0);
 		DataFile(QString &filename, unsigned int length, QObject *parent = 0);
 		~DataFile();
-		QVector<int16_t> getData(int block, int channel);
+		//QVector<int16_t> getData(int block, int channel);
+		QVector<double> getData(int block, int channel);
 
 		uint32_t getHeaderSize();
 		uint32_t getNumSamples();
