+++ recording.cpp
@@ -51,26 +51,45 @@ PlaybackRecording::PlaybackRecording(QString &filename) :
 PlaybackRecording::~PlaybackRecording() {
 }
 
-QVector<int16_t> PlaybackRecording::getData(int block, int channel) {
+//QVector<int16_t> PlaybackRecording::getData(int block, int channel) {
+	//return this->getFile().getData(block, channel);
+//}
+
+//QVector<QVector<int16_t> > PlaybackRecording::getDataBlock(int block) {
+	//QVector<int16_t> tmp(this->getFile().getBlockSize());
+	//QVector<QVector<int16_t> > data(this->getFile().getNumChannels());
+	//for (auto chan = 0; chan < this->getFile().getNumChannels(); chan++)
+		//data[chan] = this->getData(block, chan);
+	//return data;
+//}
+
+//QVector<QVector<int16_t> > PlaybackRecording::getNextDataBlock() {
+	//QVector<QVector<int16_t> > data = this->getDataBlock(this->currentBlock);
+	//this->currentBlock += 1;
+	//if (this->currentBlock == this->getFile().getNumBlocks())
+		//emit endOfPlaybackFile();
+	//return data;
+//}
+
+QVector<double> PlaybackRecording::getData(int block, int channel) {
 	return this->getFile().getData(block, channel);
 }
 
-QVector<QVector<int16_t> > PlaybackRecording::getDataBlock(int block) {
-	QVector<int16_t> tmp(this->getFile().getBlockSize());
-	QVector<QVector<int16_t> > data(this->getFile().getNumChannels());
+QVector<QVector<double> > PlaybackRecording::getDataBlock(int block) {
+	QVector<double> tmp(this->getFile().getBlockSize());
+	QVector<QVector<double> > data(this->getFile().getNumChannels());
 	for (auto chan = 0; chan < this->getFile().getNumChannels(); chan++)
 		data[chan] = this->getData(block, chan);
 	return data;
 }
 
-QVector<QVector<int16_t> > PlaybackRecording::getNextDataBlock() {
-	QVector<QVector<int16_t> > data = this->getDataBlock(this->currentBlock);
+QVector<QVector<double> > PlaybackRecording::getNextDataBlock() {
+	QVector<QVector<double> > data = this->getDataBlock(this->currentBlock);
 	this->currentBlock += 1;
 	if (this->currentBlock == this->getFile().getNumBlocks())
 		emit endOfPlaybackFile();
 	return data;
 }
-
 void PlaybackRecording::endOfPlaybackFile() {
 }
